refactor(hydra-cli): refactoring III
affects: @dzlzv/hydra-cli
