--- packages/hydra-cli/src/model/validate.ts
@@ -1,7 +1,0 @@
-import { Field } from '.'
-
-export function validateVariantField(f: Field): void {
-  if (f.isRelationType()) {
-    throw new Error(`Reference types are not supported in varaints`)
-  }
-}
